requirements.txt
fastapi
uvicorn
newspaper3k
sumy
tinysegmenter